update README and a couple more changes in the code

+++ kma_films_site/main/views.py
@@ -1,13 +1,11 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status
 from .models import Movie, Vote, Favorite, Watchlist
 from .serializers import MovieSerializer, VoteSerializer, FavoriteSerializer, WatchlistSerializer
 from django.http import HttpResponse
 
 def home(request):
     return HttpResponse("OK")
-
 
 class MovieListView(APIView):
     def get(self, request):
@@ -22,14 +20,12 @@ class MovieListView(APIView):
             return Response(serializer.data, status=201)
         return Response(serializer.errors, status=400)
 
-
 class MovieDetailView(APIView):
     def get(self, request, pk):
         try:
             movie = Movie.objects.get(id=pk)
         except Movie.DoesNotExist:
             return Response({'error': 'Movie not found'}, status=404)
-
         serializer = MovieSerializer(movie)
         return Response(serializer.data)
 
@@ -38,7 +34,6 @@ class MovieDetailView(APIView):
             movie = Movie.objects.get(id=pk)
         except Movie.DoesNotExist:
             return Response({'error': 'Movie not found'}, status=404)
-
         serializer = MovieSerializer(movie, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -50,13 +45,12 @@ class MovieDetailView(APIView):
             movie = Movie.objects.get(id=pk)
         except Movie.DoesNotExist:
             return Response({'error': 'Movie not found'}, status=404)
-
         if Favorite.objects.filter(movie=movie).exists():
-            return Response({'error': 'Cannot delete movie in favorites'}, status=400)
-
+            return Response({'error': 'Cannot delete movie that is in favorites'}, status=400)
+        if Vote.objects.filter(movie=movie).exists():
+            return Response({'error': 'Cannot delete movie that has votes'}, status=400)
         movie.delete()
         return Response({'message': 'Movie deleted'}, status=204)
-
 
 class VoteListView(APIView):
     def get(self, request):
@@ -71,7 +65,6 @@ class VoteListView(APIView):
             return Response(serializer.data, status=201)
         return Response(serializer.errors, status=400)
 
-
 class FavoriteListView(APIView):
     def get(self, request):
         favorites = Favorite.objects.all()
@@ -84,3 +77,59 @@ class FavoriteListView(APIView):
             serializer.save()
             return Response(serializer.data, status=201)
         return Response(serializer.errors, status=400)
+
+class VoteDetailView(APIView):
+    def get(self, request, pk):
+        try:
+            vote = Vote.objects.get(id=pk)
+        except Vote.DoesNotExist:
+            return Response({'error': 'Vote not found'}, status=404)
+        serializer = VoteSerializer(vote)
+        return Response(serializer.data)
+
+    def put(self, request, pk):
+        try:
+            vote = Vote.objects.get(id=pk)
+        except Vote.DoesNotExist:
+            return Response({'error': 'Vote not found'}, status=404)
+        serializer = VoteSerializer(vote, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
+
+    def delete(self, request, pk):
+        try:
+            vote = Vote.objects.get(id=pk)
+        except Vote.DoesNotExist:
+            return Response({'error': 'Vote not found'}, status=404)
+        vote.delete()
+        return Response({'message': 'Vote deleted'}, status=204)
+
+class FavoriteDetailView(APIView):
+    def get(self, request, pk):
+        try:
+            favorite = Favorite.objects.get(id=pk)
+        except Favorite.DoesNotExist:
+            return Response({'error': 'Favorite not found'}, status=404)
+        serializer = FavoriteSerializer(favorite)
+        return Response(serializer.data)
+
+    def put(self, request, pk):
+        try:
+            favorite = Favorite.objects.get(id=pk)
+        except Favorite.DoesNotExist:
+            return Response({'error': 'Favorite not found'}, status=404)
+        serializer = FavoriteSerializer(favorite, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
+
+    def delete(self, request, pk):
+        try:
+            favorite = Favorite.objects.get(id=pk)
+        except Favorite.DoesNotExist:
+            return Response({'error': 'Favorite not found'}, status=404)
+        favorite.delete()
+        return Response({'message': 'Favorite deleted'}, status=204)
